refactor: extract cell readers from DSExportRequest into separate module
